+++ lib/src/random.dart
@@ -11,7 +11,7 @@ export 'package:hashlib/src/core/hashlib_random.dart';
 @pragma('vm:prefer-inline')
 Uint8List randomBytes(
   int length, {
-  RandomGenerator generator = RandomGenerator.system,
+  RandomGenerator generator = RandomGenerator.secure,
 }) =>
     HashlibRandom(generator).nextBytes(length);
 
@@ -19,7 +19,7 @@ Uint8List randomBytes(
 @pragma('vm:prefer-inline')
 Uint32List randomNumbers(
   int length, {
-  RandomGenerator generator = RandomGenerator.system,
+  RandomGenerator generator = RandomGenerator.secure,
 }) =>
     HashlibRandom(generator).nextNumbers(length);
 
@@ -31,7 +31,7 @@ void fillRandom(
   ByteBuffer buffer, {
   int start = 0,
   int? length,
-  RandomGenerator generator = RandomGenerator.system,
+  RandomGenerator generator = RandomGenerator.secure,
 }) =>
     HashlibRandom(generator).fill(buffer, start, length);
 
@@ -42,7 +42,7 @@ void fillNumbers(
   List<int> list, {
   int start = 0,
   int? length,
-  RandomGenerator generator = RandomGenerator.system,
+  RandomGenerator generator = RandomGenerator.secure,
 }) {
   int n = length ?? list.length;
   if (n == 0) return;
@@ -65,7 +65,7 @@ String randomString(
   bool? punctuations,
   List<int>? whitelist,
   List<int>? blacklist,
-  RandomGenerator generator = RandomGenerator.system,
+  RandomGenerator generator = RandomGenerator.secure,
 }) =>
     HashlibRandom(generator).nextString(
       length,